refactor: add laminas-httphandlerrunner v2 implementation and split tests
- Creates Mezzio\Swoole\RequestHandlerRunner\V2RequestHandlerRunner, which implements Laminas\HttpHandlerRunner\RequestHandlerRunnerInterface and RequestHandlerConstantsInterface, and uses the RequestHandlerRunnerTrait to create the implementation.
- Extracts the SwooleRequestHandlerRunnerTest into an AbstractRequestHandlerRunnerTest, with two abstract methods, getRequestHandlerClass() and determineWhetherOrNotToSkipRequestHandlerRunnerTests().
  The latter will mark tests skipped if they cannot be run for the current laminas-httphandlerrunner version.
  The former is used to get the class name to use for `new()` operations.
- Modifies the SwooleRequestHandlerRunnerTest to extend the new AbstractRequestHandlerRunnerTest and fill out the abstract methods.
- Creates a V2RequestHandlerRunnerTest that extends the new AbstractRequestHandlerRunnerTest and fills out the abstract methods.
